requirements.txt
fastapi
uvicorn
onnxruntime
numpy
opencv-python
pillow
python-multipart